working but idk

+++ restaurants/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
+from django.http import Http404
 
 def signup(request):
     form = SignupForm()
@@ -42,6 +43,12 @@ def signout(request):
     logout(request)
     return redirect("signin")
 
+def no_access(request):
+    context = {
+        "msg" : 'you have no access!'
+    }
+    return render(request, 'no_access.html', context)
+
 def restaurant_list(request):
     context = {
         "restaurants":Restaurant.objects.all()
@@ -59,29 +66,38 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
-    form = RestaurantForm()
-    if request.method == "POST":
-        form = RestaurantForm(request.POST, request.FILES)
-        if form.is_valid():
-            restaurant = form.save(commit=False)
-            restaurant.owner = request.user
-            restaurant.save()
-            return redirect('restaurant-list')
+    if request.user.is_authenticated:
+        form = RestaurantForm()
+        if request.method == "POST":
+            form = RestaurantForm(request.POST, request.FILES)
+            if form.is_valid():
+                restaurant = form.save(commit=False)
+                restaurant.owner = request.user
+                restaurant.save()
+                return redirect('restaurant-list')
+    else:
+        return redirect('signin')
+
     context = {
         "form":form,
     }
     return render(request, 'create.html', context)
 
 def item_create(request, restaurant_id):
-    form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
-    if request.method == "POST":
-        form = ItemForm(request.POST)
-        if form.is_valid():
-            item = form.save(commit=False)
-            item.restaurant = restaurant
-            item.save()
-            return redirect('restaurant-detail', restaurant_id)
+
+    if request.user.is_staff or restaurant.owner == request.user:
+        form = ItemForm()
+        if request.method == "POST":
+            form = ItemForm(request.POST)
+            if form.is_valid():
+                item = form.save(commit=False)
+                item.restaurant = restaurant
+                item.save()
+                return redirect('restaurant-detail', restaurant_id)
+    else:
+        return redirect ('no_access')
+
     context = {
         "form":form,
         "restaurant": restaurant,
@@ -90,12 +106,18 @@ def item_create(request, restaurant_id):
 
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
-    form = RestaurantForm(instance=restaurant_obj)
-    if request.method == "POST":
-        form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
-        if form.is_valid():
-            form.save()
-            return redirect('restaurant-list')
+
+    if request.user.is_staff or restaurant_obj.owner == request.user:
+        restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+        form = RestaurantForm(instance=restaurant_obj)
+        if request.method == "POST":
+            form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
+            if form.is_valid():
+                form.save()
+                return redirect('restaurant-list')
+    else:
+        return redirect ('no_access')
+
     context = {
         "restaurant_obj": restaurant_obj,
         "form":form,
@@ -103,6 +125,9 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
-    restaurant_obj = Restaurant.objects.get(id=restaurant_id)
-    restaurant_obj.delete()
-    return redirect('restaurant-list')
+    if request.user.is_staff:
+        restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+        restaurant_obj.delete()
+        return redirect('restaurant-list')
+    else:
+        return redirect ('no_access')
